requirements.txt
Flask==3.1.2
psycopg==3.1.19
gunicorn==20.1.0
python-dotenv==0.21.0
requests==2.28.1